+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -115,6 +115,10 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 			new LogUpdate(log => log.ParsingVersion < new Version(1, 6, 0, 3)
 					&& log.Encounter == Encounter.HarvestTemple,
 				"Slightly improved health percentage display for Harvest Temple (16.66% per phase)"),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 7, 0, 0)
+					&& log.Encounter == Encounter.Other
+					&& log.MapId == MapIds.OldLionsCourt,
+				"Add support for Old Lion's Court"),
 			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
 			// unless the version changes more significantly, in that case it can be reset to 0.
 		};