+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -119,6 +119,10 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 					&& log.Encounter == Encounter.Other
 					&& log.MapId == MapIds.OldLionsCourt,
 				"Add support for Old Lion's Court"),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 7, 1, 0)
+					&& log.Encounter == Encounter.Other
+					&& log.MapId == MapIds.OldLionsCourt,
+				"Add support for Old Lion's Court CM"),
 			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
 			// unless the version changes more significantly, in that case it can be reset to 0.
 		};