+++ ArcdpsLogManager/CHANGELOG.md
@@ -6,7 +6,8 @@ This is the full changelog of the arcdps Log Manager.
 
 #### New features
 - Slightly improved health display for Harvest Temple (16.66% per phase), a better phase display might come in the future.
-- *Inspector*: Added an option to show times since start of log (enabled by default)
+- *EVTC Inspector*: Added an option to show times since start of log (enabled by default)
+- *EVTC Inspector*: Added more log information (map id, game build, ...) to statistics tab
 
 #### Changes
 - Linux only (already worked on Windows): Clicking on the filename in the log detail panel now selects the file (requires dbus).