+++ ArcdpsLogManager/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the full changelog of the arcdps Log Manager.
 
+## Log Manager v1.7.1
+#### New features
+- Added support for Old Lion's Court CM
+
 ## Log Manager v1.7
 
 #### New features